refactor: reduced the sonar score for new features
