+++ api/src/test/java/org/lfenergy/compas/scl/data/rest/api/history/ErrorResponseDtoTest.java
@@ -4,6 +4,7 @@
 package org.lfenergy.compas.scl.data.rest.api.history;
 
 import org.junit.jupiter.api.Test;
+import org.lfenergy.compas.scl.data.rest.api.errors.ErrorResponseDto;
 
 import java.time.OffsetDateTime;
